+++ CHANGES.md
@@ -7,6 +7,7 @@
 -   Capture all errors logged by gdal when opening a file fails (#495).
 -   Add support to read and write ".gpkg.zip" (GDAL >= 3.7), ".shp.zip", and ".shz"
     files (#527).
+-   Compatibility with the string dtype in the upcoming pandas 3.0 release (#493).
 
 ### Bug fixes